+++ src/ForeignPacketsReceiver.h
@@ -13,12 +13,12 @@
 
 class ForeignPacketsReceiver {
  public:
-  explicit ForeignPacketsReceiver(std::vector<std::string> wlans,std::vector<uint16_t> openhd_radio_ports);
+  explicit ForeignPacketsReceiver(std::vector<std::string> wlans,std::vector<int> openhd_radio_ports);
  private:
   void on_foreign_packet(uint8_t wlan_idx, const pcap_pkthdr &hdr, const uint8_t *pkt);
   void m_loop();
   std::unique_ptr<MultiRxPcapReceiver> m_receiver;
-  std::vector<uint16_t> m_openhd_radio_ports;
+  std::vector<int> m_openhd_radio_ports;
   std::unique_ptr<std::thread> m_thread;
 };
